Fixed sign conversion errors.

+++ Source/UI/SDTableViewSectionController.m
@@ -37,7 +37,7 @@
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[section];
+    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[(NSUInteger)section];
     if ([sectionController respondsToSelector:@selector(numberOfRowsForSectionController:)])
     {
         return [sectionController numberOfRowsForSectionController:self];
@@ -47,7 +47,7 @@
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSInteger section = indexPath.section;
+    NSUInteger section = (NSUInteger)indexPath.section;
     NSInteger row = indexPath.row;
     id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[section];
     if ([sectionController respondsToSelector:@selector(sectionController:cellForRow:)])
@@ -76,7 +76,7 @@
             strongTableView.delegate = self;
             strongTableView.dataSource = self;
             
-            NSInteger sectionCount = self.sectionControllers.count;
+            NSInteger sectionCount = (NSInteger)self.sectionControllers.count;
             return sectionCount;
         }
     }
@@ -86,7 +86,7 @@
 
 - (NSString *)tableView:(UITableView *)tableView titleForHeaderInSection:(NSInteger)section
 {
-    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[section];
+    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[(NSUInteger)section];
     if ([sectionController respondsToSelector:@selector(sectionControllerTitleForHeader:)])
     {
         return [sectionController sectionControllerTitleForHeader:self];
@@ -102,7 +102,7 @@
 {
     NSInteger section = indexPath.section;
     NSInteger row = indexPath.row;
-    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[section];
+    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[(NSUInteger)section];
     if ([sectionController respondsToSelector:@selector(sectionController:didSelectRow:)])
     {
         [sectionController sectionController:self didSelectRow:row];
@@ -113,7 +113,7 @@
 {
     NSInteger section = indexPath.section;
     NSInteger row = indexPath.row;
-    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[section];
+    id<SDTableViewSectionDelegate>sectionController = self.sectionControllers[(NSUInteger)section];
     if ([sectionController respondsToSelector:@selector(sectionController:heightForRow:)])
     {
         return [sectionController sectionController:self heightForRow:row];
